+++ murrelet_livecode/src/lib.rs
@@ -6,3 +6,4 @@ pub mod nestedit;
 pub mod state;
 pub mod types;
 pub mod unitcells;
+pub mod lazy;
